+++ src/features/Home/components/CategorySection.tsx
