--- src/pages/Home.tsx
@@ -1,17 +1,0 @@
-import CategorySection from '@components/HomePage/CategorySection';
-import PromoBanner from '@components/HomePage/PromoBanner';
-import RankingSection from '@components/HomePage/RankingSection';
-import FreiendSelector from '@components/HomePage/FriendSelector';
-
-const Home = () => {
-  return (
-    <>
-      <FreiendSelector />
-      <CategorySection />
-      <PromoBanner />
-      <RankingSection />
-    </>
-  );
-};
-
-export default Home;
