+++ tsconfig.app.json
@@ -13,7 +13,8 @@
       "@common/*": ["src/components/common/*"],
       "@contexts/*": ["src/contexts/*"],
       "@schemas/*": ["src/schemas/*"],
-      "@apis/*": ["src/apis/*"]
+      "@apis/*": ["src/apis/*"],
+      "@features/*": ["src/features/*"]
     },
     "types": ["node"],
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
